+++ src/main/java/jact/lagaltproject/services/project/ProjectService.java
@@ -3,5 +3,5 @@ package jact.lagaltproject.services.project;
 import jact.lagaltproject.models.Project;
 import jact.lagaltproject.services.CrudService;
 
-public interface ProjectService extends CrudService<Project, Integer> {
+public interface ProjectService extends CrudService<Project, Long> {
 }
